Backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic[email]
sqlalchemy
databases
email-validator
